+++ gui/src/core/behaviors/ForkBehavior.cpp
@@ -15,16 +15,16 @@ void ForkBehavior::_spawnParticles()
 {
     for (int i = 0; i < PARTICLE_COUNT; i++) {
         float angle = (static_cast<float>(i) / PARTICLE_COUNT) * 2.0f * PI +
-                      (static_cast<float>(rand()) / RAND_MAX) * 0.4f;
-        float speed = 0.9f + (static_cast<float>(rand()) / RAND_MAX) * 0.85f;
+                      (static_cast<float>(rand()) / static_cast<float>(RAND_MAX)) * 0.4f;
+        float speed = 0.9f + (static_cast<float>(rand()) / static_cast<float>(RAND_MAX)) * 0.85f;
 
         Particle p;
         p.pos = {_visual.pos.x, _visual.pos.y + 0.1f, _visual.pos.z};
         p.vel.x = std::cos(angle) * speed;
-        p.vel.y = 0.6f + (static_cast<float>(rand()) / RAND_MAX) * 0.5f;
+        p.vel.y = 0.6f + (static_cast<float>(rand()) / static_cast<float>(RAND_MAX)) * 0.5f;
         p.vel.z = std::sin(angle) * speed;
         p.color = {180, 255, 120, 255};  // vivid green
-        p.size = 0.06f + (static_cast<float>(rand()) / RAND_MAX) * 0.05f;
+        p.size = 0.06f + (static_cast<float>(rand()) / static_cast<float>(RAND_MAX)) * 0.05f;
         p.alpha = 1.0f;
         p.delay = 0.0f;
         p.active = true;
